Make id_valid field of bop_record_t optional (through macros)
Since Object.use/.promise may not be needed (which is the only time that id_valid is actually needed -- all other instances have id_valid == true) this will decrease the size of the struct.
This in turn allows for more to be packed into the shared memory area used for bookkeeping.

Also add a simple function defined in object_monitor.h that checks for validity of a bop_record. Returns false for null records. Tests the struct field when HAVE_USE_PROMISE is defined (and hence id_valid field exits. When not define defined as record == null ? false : true